+++ changelog.md
@@ -1,5 +1,3 @@
 # Changelog 0.6.1
 
-## Bugs fixed
-
-- Fixed unexpected conflict error after merge
+- Permission setup mention users or roles silently